ir: refactor metadata enabling out of `New` function
Not functional changes.
